+++ modules/hdf5/src/cpp/h5Save.cpp
@@ -16,10 +16,12 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
+#include <ctime>
 #define H5_BUILT_AS_DYNAMIC_LIB
 #include <hdf5.h>
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/path.hpp>
+#include <boost/algorithm/string.hpp>
 #include "h5Save.hpp"
 #include "IsValidVariableName.hpp"
 #include "characters_encoding.hpp"
@@ -30,6 +32,79 @@
 #include "h5SaveVariable.hpp"
 //=============================================================================
 namespace Nelson {
+//=============================================================================
+static std::wstring
+createHeader()
+{
+    std::wstring header = std::wstring(NELSON_HEADER);
+    time_t _tm = time(nullptr);
+    struct tm* curtime = localtime(&_tm);
+    std::wstring timestr = utf8_to_wstring(asctime(curtime));
+    boost::algorithm::replace_last(timestr, L"\n", L"");
+    return header + std::wstring(L" on ") + timestr;
+}
+//=============================================================================
+static hid_t
+createNh5FileWithHeader(const std::wstring& filename, const std::wstring& header)
+{
+    char* header_offset = nullptr;
+    try {
+        header_offset = new char[8];
+        memset(header_offset, ' ', (size_t)8 * sizeof(char));
+    } catch (std::bad_alloc&) {
+        return H5I_INVALID_HID;
+    }
+    char* header_saturated = nullptr;
+    try {
+        header_saturated = new char[128];
+        memset(header_saturated, ' ', (size_t)128 * sizeof(char));
+    } catch (std::bad_alloc&) {
+        delete[] header_offset;
+        return H5I_INVALID_HID;
+    }
+
+    std::string utf_filename = wstring_to_utf8(filename);
+    hid_t plist_id = H5Pcreate(H5P_FILE_CREATE);
+    H5Pset_userblock(plist_id, 512);
+    hid_t plist_ap = H5Pcreate(H5P_FILE_ACCESS);
+#if H5_VERSION_GE(1, 10, 2)
+    H5Pset_libver_bounds(plist_ap, H5F_LIBVER_EARLIEST, H5F_LIBVER_V18);
+#endif
+    hid_t fid = H5Fcreate(utf_filename.c_str(), H5F_ACC_TRUNC, plist_id, plist_ap);
+    H5Fclose(fid);
+    H5Pclose(plist_id);
+    if (fid < 0) {
+        delete[] header_offset;
+        delete[] header_saturated;
+        return fid;
+    }
+    FILE* fp = fopen(utf_filename.c_str(), "r+b");
+    if (fp == nullptr) {
+        delete[] header_offset;
+        delete[] header_saturated;
+        return H5I_INVALID_HID;
+    }
+    (void)fseek(fp, 0, SEEK_SET);
+    size_t len = snprintf(header_saturated, 116, "%s", wstring_to_utf8(header).c_str());
+    if (len >= 116) {
+        header_saturated[115] = '\0';
+    }
+
+    int16 endian = NELSON_HEADER_ENDIAN;
+    int16 version = NELSON_HEADER_VERSION;
+
+    fwrite(header_saturated, 1, 116, fp);
+    fwrite(header_offset, 1, 8, fp);
+    fwrite(&version, 2, 1, fp);
+    fwrite(&endian, 2, 1, fp);
+    fclose(fp);
+
+    fid = H5Fopen(utf_filename.c_str(), H5F_ACC_RDWR, plist_ap);
+    H5Pclose(plist_ap);
+    delete[] header_offset;
+    delete[] header_saturated;
+    return fid;
+}
 //=============================================================================
 void
 h5Save(Evaluator* eval, const std::wstring& filename, wstringVector names, bool append,
@@ -64,8 +139,7 @@ h5Save(Evaluator* eval, const std::wstring& filename, wstringVector names, bool 
         fileExistPreviously = false;
     }
     if (!fileExistPreviously) {
-        fid = H5Fcreate(wstring_to_utf8(hdf5_filename.wstring()).c_str(), H5F_ACC_TRUNC,
-            H5P_DEFAULT, H5P_DEFAULT);
+        fid = createNh5FileWithHeader(hdf5_filename.wstring(), createHeader());
     } else {
         if (!H5Fis_hdf5(wstring_to_utf8(hdf5_filename.wstring()).c_str()))
             Error(_W("HDF5 format file expected."));
@@ -81,8 +155,7 @@ h5Save(Evaluator* eval, const std::wstring& filename, wstringVector names, bool 
                     Error(_W("Cannot replace file"));
                     boost::system::error_code error_code = e.code();
                 }
-                fid = H5Fcreate(wstring_to_utf8(hdf5_filename.wstring()).c_str(), H5F_ACC_TRUNC,
-                    H5P_DEFAULT, H5P_DEFAULT);
+                fid = createNh5FileWithHeader(hdf5_filename.wstring(), createHeader());
                 updateNelsonH5Header(fid);
             }
         }